requirements.txt
Django==4.2.1
psycopg2==2.9.6
graphene-django==3.1.1
